+++ src/lib/sensors.ts
@@ -5,8 +5,59 @@ import {
   CadenceMeter,
   // @ts-ignore
 } from 'react-native-cycling-sensors';
+import {dataBase} from '../database';
+import SensorModel from '../database/model/sensorModel';
 
 export const ble = new BleSensors();
 export const powerMeter = new PowerMeter();
 export const heartRateMonitor = new HeartRateMonitor();
 export const cadenceMeter = new CadenceMeter();
+
+export function bleCharacteristicToIconName(char: string): string {
+  switch (char) {
+    case 'Battery':
+      // Battery Service
+      return 'battery-bluetooth-variant';
+    case 'HeartRate':
+      // Heart rate service
+      return 'heart-pulse';
+    case 'CyclingPower':
+      // Cycling Power
+      return 'lightning-bolt';
+    case 'CyclingSpeedAndCadence':
+      // Cycling Speed And Cadence
+      return 'unicycle';
+    default:
+      return 'ab-testing';
+  }
+}
+
+export async function createSensor(
+  name: string,
+  address: string,
+  services: string[],
+  type: string = 'bluetooth',
+): Promise<SensorModel> {
+  console.log('creating...');
+  return dataBase.write(async () => {
+    return dataBase
+      .get<SensorModel>('sensor')
+      .create(data => {
+        data.name = name;
+        data.type = type;
+        data.address = address;
+        data.sensorType = services;
+        data.createdAt = new Date().getTime();
+        console.log('creating sensor', data);
+        return data;
+      })
+      .catch(err => {
+        console.log('got error creating sensor: ', err);
+        throw err;
+      });
+  });
+}
+
+export function getAllSensors() {
+  return dataBase.get<SensorModel>('sensor').query().fetch();
+}
